pyproject.toml
[tool.poetry]
name = "data-analysis"
version = "0.1.0"
description = "Structured data analysis example"
authors = ["Fernando Robledo <fernando.robledoruiz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10,<=3.13"
crewai = "^0.22.5"
pandas = "^2.2.1"
crewai-tools = "^0.1.0"
langchain-google-genai = "^0.0.11"
langchain-groq = "^0.0.1"
langchain-anthropic = "^0.1.4"
pandasai = "^2.0.21"
langchain-experimental = "^0.0.55"
tabulate = "^0.9.0"
streamlit = "^1.32.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
